removed SharedTableCustomDynamic and SharedTableCustomStatic strategies, consolidated documentation
